refactor(runtime/session): split tests to sibling tests_session.rs (#1266)
Inline test module in session.rs is moved into a sibling tests_session.rs
following the existing runtime convention (already has tests_mod.rs).

- src/runtime/session.rs: 505 → 375 LOC (under 400-line rule)
- src/runtime/tests_session.rs: new (~127 LOC)
- src/runtime/mod.rs: register mod tests_session

Pure refactor: zero logic changes, all 7 tests pass.

Closes 16/70 of #1266 (this loop iteration), 25 PRs cumulative since #1285.
